Fix character switch + Removed wait from main loop which could caused flickering
